GH-382: [refactor] Rename package for forgot password events to follow naming conventions
