+++ Microservices/email-service/src/test/java/app/sportahub/emailservice/service/kafka/consumer/OrchestrationServiceTest.java
@@ -1,6 +1,6 @@
 package app.sportahub.emailservice.service.kafka.consumer;
 
-import app.sportahub.kafkevents.forgotPassword.ForgotPasswordSendEmailEvent;
+import app.sportahub.kafkevents.forgotpassword.ForgotPasswordSendEmailEvent;
 import app.sportahub.kafkevents.BaseEvent;
 import app.sportahub.emailservice.service.EmailService;
 import jakarta.mail.MessagingException;
