+++ Microservices/event-service/src/main/java/app/sportahub/eventservice/service/kafka/consumer/OrchestrationServiceConsumerImpl.java
@@ -3,9 +3,9 @@ package app.sportahub.eventservice.service.kafka.consumer;
 import app.sportahub.eventservice.model.event.Event;
 import app.sportahub.eventservice.repository.event.EventRepository;
 import app.sportahub.kafkevents.BaseEvent;
-import app.sportahub.kafkevents.JoinedSportEventEvent.JoinedEventsByUserEvent;
-import app.sportahub.kafkevents.JoinedSportEventEvent.JoinedEventsByUserFetchEvent;
-import app.sportahub.kafkevents.JoinedSportEventEvent.JoinedEventsByUserFetchedEvent;
+import app.sportahub.kafkevents.joinsporteventevent.JoinedEventsByUserEvent;
+import app.sportahub.kafkevents.joinsporteventevent.JoinedEventsByUserFetchEvent;
+import app.sportahub.kafkevents.joinsporteventevent.JoinedEventsByUserFetchedEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.producer.ProducerRecord;
