+++ Microservices/orchestration-service/src/main/java/app/sportahub/orchestrationservice/service/consumer/UserServiceConsumer.java
@@ -1,7 +1,7 @@
 package app.sportahub.orchestrationservice.service.consumer;
 
-import app.sportahub.kafkevents.JoinedSportEventEvent.JoinedEventsByUserRequestEvent;
-import app.sportahub.kafkevents.forgotPassword.ForgotPasswordRequestedEvent;
+import app.sportahub.kafkevents.joinsporteventevent.JoinedEventsByUserRequestEvent;
+import app.sportahub.kafkevents.forgotpassword.ForgotPasswordRequestedEvent;
 
 public interface UserServiceConsumer {
     public void listenForForgotPasswordRequestedEvent(ForgotPasswordRequestedEvent forgotPasswordRequestedEvent);
