+++ Microservices/user-service/src/main/java/app/sportahub/userservice/config/kafka/KafkaProducerConfig.java
@@ -1,12 +1,11 @@
 package app.sportahub.userservice.config.kafka;
 
-import app.sportahub.kafkevents.JoinedSportEventEvent.JoinedEventsByUserEvent;
+import app.sportahub.kafkevents.joinsporteventevent.JoinedEventsByUserEvent;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.common.serialization.StringSerializer;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.annotation.EnableKafka;
 import org.springframework.kafka.core.ConsumerFactory;
 import org.springframework.kafka.core.DefaultKafkaProducerFactory;
 import org.springframework.kafka.core.KafkaTemplate;
