+++ libs/kafka-events/src/main/java/app/sportahub/kafkevents/joinsporteventevent/JoinedEventsByUserRequestEvent.java
@@ -1,15 +1,15 @@
-package app.sportahub.kafkevents.JoinedSportEventEvent;
+package app.sportahub.kafkevents.joinsporteventevent;
 
 import app.sportahub.kafkevents.BaseEvent;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class JoinedEventsByUserFetchEvent {
+public class JoinedEventsByUserRequestEvent {
     private final BaseEvent baseEvent;
     private final String userId;
 
     @JsonCreator
-    public JoinedEventsByUserFetchEvent(
+    public JoinedEventsByUserRequestEvent(
             @JsonProperty("baseEvent") BaseEvent baseEvent,
             @JsonProperty("userId") String userId)
     {
